评比部分后台管理

+++ src/main/java/com/aifeng/service/RatingResultService.java
@@ -48,6 +48,8 @@ public class RatingResultService {
     @Transactional
     public void delRatingResult(RatingObj ratingObj) {
         RatingResult ratingResult = ratingResultRepository.findByRatingObj(ratingObj);
-        ratingResultRepository.delete(ratingResult);
+        if(ratingResult != null) {
+            ratingResultRepository.delete(ratingResult);
+        }
     }
 }
